refactor: decompose skill-lab.ts into focused modules
Split the 1230-line skill-lab.ts into 5 focused modules:
- behavior-benchmark.ts: behavior benchmark logic
- improve.ts: skill improvement logic
- reporting.ts: reporting and output formatting
- scaffold.ts: skill scaffolding
- trigger-benchmark.ts: trigger benchmark logic

All 27 skill-lab tests pass. Public API unchanged.

Closes #134
